Allow manually trigger PaC builds
When component is provisioned for PaC, we are able to re-trigger push
pipeline (which wouldn't be otherwise possible) by adding request
annotation to component:
build.appstudio.openshift.io/request: trigger-pac-build

It will update incomings in repository, and create secret which is used
for trigger. And run push pipeline for a component.

STONEBLD-1438

Signed-off-by: Robert Cerven <rcerven@redhat.com>