Tekton v1: log when converting v1beta1 to v1
To have some easily searchable stats about the usage of the deprecated
v1beta1 APIs, log a message any time we retrieve a v1beta1 Pipeline from
a pipelineRef.

Note: this is relevant only for generating PaC configs, not for the
simple build workflow. For a simple build, we never retrieve the whole
Pipeline object. We generate a PipelineRun that includes the pipelineRef
and send the PipelineRun to the cluster, Tekton figures out the rest.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>