Add script for setting appstudio_workspace params
STONEBLD-2008

For newly created Repository objects, build-service will be setting the
appstudio_workspace param. For Repository objects created before this
feature was deployed, the appstudio_workspace param will be missing.

Add a script that automatically sets the appstudio_workspace param for
all Repositories that don't have it set correctly yet.

The script requires python>=3.8 and depends on kubectl or oc (and
nothing else).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>