fix(rhtapbugs=975): use source_url instead of repo_url for git-url
The repo_url variable in pull request jobs points to the target repo.
The git-url parameter of build pipelineRuns is passed to the integration
service.  If users attempt to clone this URL with the revision from the
'revision' parameter (which is the revision of the source URL) an error
will occur.  Tekton also has a 'source_url' variable that points to the
source repo in pull request jobs, which fixes the problem.  In push jobs
'source_repo' defaults to the same value as 'repo_url'.  Therefore
defaulting to 'source_repo' should not cause any other bugs'

Signed-off-by: Ryan Cole <rcyoalne@gmail.com>