Enable the customization of nudged files
Fixes KFLUXBUGS-1114

To prevent us from having to continuously modify the set of files that
can have nudged references, we will define the default set of files and
let users customize the nudges based on pipeline parameters. There is no
intention to add default values or to expose this parameter to users in
the PipelineDefinition pushed out to repositories.

Signed-off-by: arewm <arewm@users.noreply.github.com>