RHTAPSRE-469: Webhook configuration based on repo url
Allow to configure the webhook target url based on the repository
url of the component by providing a config file.

Configuring the webhook target url using environment variable has
precedence.

In case that no config file or environment variable for configuring
the webhook target are provided, the PAC route will be used (same
as before this change).

Signed-off-by: gbenhaim <gbenhaim@redhat.com>