RHTAPSRE-467: Update application-service version
In order to get the changes introduced in
https://github.com/redhat-appstudio/application-service/pull/450,
update the application service version.

In the new version some of the function moved between modules, so fixing
the import paths for those.

In addition, when of the assertions looked for an error message that
was changed, so update it as well.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>