getGitProvider: make the heuristic looser
Previously, the function guessed the provider based on a specific
segment in the hostname in the repo url.

For a host like cooler.gitlab.my-company.com, the returned provider
would be (depending on whether the repo url is git@ -style or not):

    git@cooler.gitlab.my-company.com:org/repo
        ^^^^^^

    https://cooler.gitlab.my-company.com/org/repo.git
                          ^^^^^^^^^^

Fix the heuristic and make it much looser - now it just searches the
hostname for substrings corresponding to the known git providers.

Note: in the highly unusual case where a hostname contains substrings
corresponding to more than one provider, the precedence is github >
gitlab > bitbucket. (Making the precedence deterministic is the reason
why allowedGitProviders had to change from a map to a slice).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>