manually update konflux references
Enterprise Contract checks are currently failing due to outdated task
versions.
There is a Mintmaker PR open, but it does not include the latest fix for
CycloneDX version and the PR will not get updated until Saturday.

Manually update to the references provided by Mintmaker, but also
include the newer buildah task version which has the fix for CycloneDX
version.