Add support for canceling in-progress pipeline runs
With PAC v0.32.0, support was added to cancel in-progress pipeline runs,
see:
https://pipelinesascode.com/docs/guide/running/#cancelling-in-progress-pipelineruns.
Even if this version of PAC is not deployed, by starting to add the
annotation, users can take advantage of the functionality when it is
deployed.

While the annotation is added to both push and PR events, it is only
enabled by default for PR events. This enables all push events to run to
completion if multiple pushes happen in series.

ref: https://github.com/openshift-pipelines/pipelines-as-code/releases/tag/v0.32.0

Signed-off-by: arewm <arewm@users.noreply.github.com>