Fix appstudio-pipeline Service Account Secrets sync for newly created Components (#436)
Handle the situation when both the Component and the Image Repository were created at the same time and Image Controller operator hasn't yet set the owner reference for the Image Repository. Even in such case, the push secret should not be treated as common and should not be synced to other Components build pipeline Service Account.