fix: when nudging, have Renovate be more patient with GitLab (#466)
GitLab has a long-standing issue where the API refuses attempts to set
"merge when pipeline succeeds" even when the MR is ready for it[1].

Renovate already does some retrying with exponential backoff[2], but
the default behavior is to only wait for up to 7.5 seconds. This leads
to nudging MRs very often failing to automerge, especially at times of
high GitLab load.

By setting a higher base delay value and a higher number of retries we
can ensure automerge works a lot more reliably.

KONFLUX-9373

[1] https://gitlab.com/gitlab-org/gitlab/-/issues/353984
[2] https://github.com/renovatebot/renovate/blob/88768303ee7f246da3337c5d34cace8c53f1284e/lib/modules/platform/gitlab/index.ts#L694-L714